requirements.txt
discord.py
pymongo
requests
paypalrestsdk
